refactor: move payment logic to BookingPaymentService
- Created dedicated payment service
- Refactored payment intent creation
- Centralized webhook processing
- Improved payment confirmation logic
